CR-660 WebChat split out - more to follow
