+++ tests/unit/test_handlers.py
@@ -1,5 +1,4 @@
 import json
-import datetime
 
 from unittest import mock
 from urllib.parse import urlsplit, parse_qs
@@ -8,13 +7,11 @@ from aiohttp.client_exceptions import ClientConnectionError
 from aiohttp.test_utils import unittest_run_loop
 from aioresponses import aioresponses
 
-from app import (BAD_CODE_MSG, INVALID_CODE_MSG, WEBCHAT_MISSING_QUERY_MSG,
-                 WEBCHAT_MISSING_COUNTRY_MSG, WEBCHAT_MISSING_NAME_MSG,
+from app import (BAD_CODE_MSG, INVALID_CODE_MSG,
                  POSTCODE_INVALID_MSG, BAD_CODE_MSG_CY, INVALID_CODE_MSG_CY,
-                 WEBCHAT_MISSING_QUERY_MSG_CY, WEBCHAT_MISSING_COUNTRY_MSG_CY,
-                 WEBCHAT_MISSING_NAME_MSG_CY, POSTCODE_INVALID_MSG_CY)
+                 POSTCODE_INVALID_MSG_CY)
 from app.exceptions import InactiveCaseError, InvalidEqPayLoad
-from app.handlers import IndexEN, IndexCY, IndexNI, WebChat
+from app.handlers import IndexEN, IndexCY, IndexNI
 
 from . import RHTestCase, build_eq_raises, skip_encrypt
 
@@ -2204,654 +2201,6 @@ class TestHandlers(RHTestCase):
             self.assertIn("Mae\\'n flin gennym, aeth rhywbeth o\\'i le",
                           contents)
             self.assertIn(self.ons_logo_cy, contents)
-
-    def test_check_open_weekday_open_census_weekend(self):
-        mocked_now = datetime.datetime(2019, 10, 12, 9, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-            self.assertTrue(WebChat.check_open())
-
-    def test_check_open_weekday_closed_census_weekend(self):
-        mocked_now = datetime.datetime(2019, 10, 13, 6, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-            self.assertFalse(WebChat.check_open())
-
-    def test_check_open_weekday_open(self):
-        mocked_now = datetime.datetime(2019, 6, 17, 9, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-            self.assertTrue(WebChat.check_open())
-
-    def test_check_open_weekday_closed(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 19, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-            self.assertFalse(WebChat.check_open())
-
-    def test_check_open_saturday_open(self):
-        mocked_now = datetime.datetime(2019, 6, 15, 9, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-            self.assertTrue(WebChat.check_open())
-
-    def test_check_open_saturday_closed(self):
-        mocked_now = datetime.datetime(2019, 6, 15, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-            self.assertFalse(WebChat.check_open())
-
-    def test_check_open_sunday_closed(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-            self.assertFalse(WebChat.check_open())
-
-    @unittest_run_loop
-    async def test_get_webchat_open_en(self):
-        mocked_now = datetime.datetime(2019, 6, 15, 9, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            response = await self.client.request('GET', self.get_webchat_en)
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.ons_logo_en, contents)
-            self.assertIn('Enter your name', contents)
-            self.assertEqual(contents.count('radio__input'), 10)
-            self.assertIn('type="submit"', contents)
-
-    @unittest_run_loop
-    async def test_get_webchat_open_cy(self):
-        mocked_now = datetime.datetime(2019, 6, 15, 9, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            response = await self.client.request('GET', self.get_webchat_cy)
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.ons_logo_cy, contents)
-            self.assertIn('Nodwch eich enw', contents)
-            self.assertEqual(contents.count('radio__input'), 10)
-            self.assertIn('type="submit"', contents)
-
-    @unittest_run_loop
-    async def test_get_webchat_open_ni(self):
-        mocked_now = datetime.datetime(2019, 6, 15, 9, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            response = await self.client.request('GET', self.get_webchat_ni)
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.nisra_logo, contents)
-            self.assertIn('Enter your name', contents)
-            self.assertEqual(contents.count('radio__input'), 10)
-            self.assertIn('type="submit"', contents)
-
-    @unittest_run_loop
-    async def test_get_webchat_not_open_200_en(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url, status=200)
-
-                response = await self.client.request('GET',
-                                                     self.get_webchat_en)
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.ons_logo_en, contents)
-            self.assertIn('Bank Holidays', contents)
-
-    @unittest_run_loop
-    async def test_get_webchat_not_open_200_cy(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url, status=200)
-
-                response = await self.client.request('GET',
-                                                     self.get_webchat_cy)
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.ons_logo_cy, contents)
-            self.assertIn('Gwyliau Banc', contents)
-
-    @unittest_run_loop
-    async def test_get_webchat_not_open_200_ni(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url, status=200)
-
-                response = await self.client.request('GET',
-                                                     self.get_webchat_ni)
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.nisra_logo, contents)
-            self.assertIn('Bank Holidays', contents)
-
-    @unittest_run_loop
-    async def test_get_webchat_not_open_clientconnectionerror_en(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url,
-                           exception=ClientConnectionError('Failed'))
-
-                with self.assertLogs('respondent-home', 'ERROR') as cm:
-                    response = await self.client.request(
-                        'GET', self.get_webchat_en)
-                self.assertLogEvent(cm, 'client failed to connect')
-                self.assertLogEvent(cm, 'failed to send webchat closed')
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.ons_logo_en, contents)
-            self.assertIn('Bank Holidays', contents)
-
-    @unittest_run_loop
-    async def test_get_webchat_not_open_clientconnectionerror_cy(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url,
-                           exception=ClientConnectionError('Failed'))
-
-                with self.assertLogs('respondent-home', 'ERROR') as cm:
-                    response = await self.client.request(
-                        'GET', self.get_webchat_cy)
-                self.assertLogEvent(cm, 'client failed to connect')
-                self.assertLogEvent(cm, 'failed to send webchat closed')
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.ons_logo_cy, contents)
-            self.assertIn('Gwyliau Banc', contents)
-
-    @unittest_run_loop
-    async def test_get_webchat_not_open_clientconnectionerror_ni(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url,
-                           exception=ClientConnectionError('Failed'))
-
-                with self.assertLogs('respondent-home', 'ERROR') as cm:
-                    response = await self.client.request(
-                        'GET', self.get_webchat_ni)
-                self.assertLogEvent(cm, 'client failed to connect')
-                self.assertLogEvent(cm, 'failed to send webchat closed')
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.nisra_logo, contents)
-            self.assertIn('Bank Holidays', contents)
-
-    @unittest_run_loop
-    async def test_get_webchat_not_open_500_en(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url, status=500)
-
-                with self.assertLogs('respondent-home', 'ERROR') as cm:
-                    response = await self.client.request(
-                        'GET', self.get_webchat_en)
-
-                self.assertLogEvent(cm, 'failed to send webchat closed')
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.ons_logo_en, contents)
-            self.assertIn('Bank Holidays', contents)
-
-    @unittest_run_loop
-    async def test_get_webchat_not_open_500_cy(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url, status=500)
-
-                with self.assertLogs('respondent-home', 'ERROR') as cm:
-                    response = await self.client.request(
-                        'GET', self.get_webchat_cy)
-
-                self.assertLogEvent(cm, 'failed to send webchat closed')
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.ons_logo_cy, contents)
-            self.assertIn('Gwyliau Banc', contents)
-
-    @unittest_run_loop
-    async def test_get_webchat_not_open_500_ni(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url, status=500)
-
-                with self.assertLogs('respondent-home', 'ERROR') as cm:
-                    response = await self.client.request(
-                        'GET', self.get_webchat_ni)
-
-                self.assertLogEvent(cm, 'failed to send webchat closed')
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.nisra_logo, contents)
-            self.assertIn('Bank Holidays', contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_incomplete_query_en(self):
-        form_data = self.webchat_form_data.copy()
-        del form_data['query']
-
-        with self.assertLogs('respondent-home', 'INFO') as cm:
-            response = await self.client.request('POST',
-                                                 self.post_webchat_en,
-                                                 data=form_data)
-        self.assertLogEvent(cm, 'form submission error')
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.ons_logo_en, contents)
-        self.assertMessagePanel(WEBCHAT_MISSING_QUERY_MSG, contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_incomplete_query_cy(self):
-        form_data = self.webchat_form_data.copy()
-        del form_data['query']
-
-        with self.assertLogs('respondent-home', 'INFO') as cm:
-            response = await self.client.request('POST',
-                                                 self.post_webchat_cy,
-                                                 data=form_data)
-        self.assertLogEvent(cm, 'form submission error')
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.ons_logo_cy, contents)
-        self.assertMessagePanel(WEBCHAT_MISSING_QUERY_MSG_CY, contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_incomplete_query_ni(self):
-        form_data = self.webchat_form_data.copy()
-        del form_data['query']
-
-        with self.assertLogs('respondent-home', 'INFO') as cm:
-            response = await self.client.request('POST',
-                                                 self.post_webchat_ni,
-                                                 data=form_data)
-        self.assertLogEvent(cm, 'form submission error')
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.nisra_logo, contents)
-        self.assertMessagePanel(WEBCHAT_MISSING_QUERY_MSG, contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_incomplete_country_en(self):
-        form_data = self.webchat_form_data.copy()
-        del form_data['country']
-
-        with self.assertLogs('respondent-home', 'INFO') as cm:
-            response = await self.client.request('POST',
-                                                 self.post_webchat_en,
-                                                 data=form_data)
-        self.assertLogEvent(cm, 'form submission error')
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.ons_logo_en, contents)
-        self.assertMessagePanel(WEBCHAT_MISSING_COUNTRY_MSG, contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_incomplete_country_cy(self):
-        form_data = self.webchat_form_data.copy()
-        del form_data['country']
-
-        with self.assertLogs('respondent-home', 'INFO') as cm:
-            response = await self.client.request('POST',
-                                                 self.post_webchat_cy,
-                                                 data=form_data)
-        self.assertLogEvent(cm, 'form submission error')
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.ons_logo_cy, contents)
-        self.assertMessagePanel(WEBCHAT_MISSING_COUNTRY_MSG_CY, contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_incomplete_country_ni(self):
-        form_data = self.webchat_form_data.copy()
-        del form_data['country']
-
-        with self.assertLogs('respondent-home', 'INFO') as cm:
-            response = await self.client.request('POST',
-                                                 self.post_webchat_ni,
-                                                 data=form_data)
-        self.assertLogEvent(cm, 'form submission error')
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.nisra_logo, contents)
-        self.assertMessagePanel(WEBCHAT_MISSING_COUNTRY_MSG, contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_incomplete_name_en(self):
-        form_data = self.webchat_form_data.copy()
-        del form_data['screen_name']
-
-        with self.assertLogs('respondent-home', 'INFO') as cm:
-            response = await self.client.request('POST',
-                                                 self.post_webchat_en,
-                                                 data=form_data)
-        self.assertLogEvent(cm, 'form submission error')
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.ons_logo_en, contents)
-        self.assertMessagePanel(WEBCHAT_MISSING_NAME_MSG, contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_incomplete_name_cy(self):
-        form_data = self.webchat_form_data.copy()
-        del form_data['screen_name']
-
-        with self.assertLogs('respondent-home', 'INFO') as cm:
-            response = await self.client.request('POST',
-                                                 self.post_webchat_cy,
-                                                 data=form_data)
-        self.assertLogEvent(cm, 'form submission error')
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.ons_logo_cy, contents)
-        self.assertMessagePanel(WEBCHAT_MISSING_NAME_MSG_CY, contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_incomplete_name_ni(self):
-        form_data = self.webchat_form_data.copy()
-        del form_data['screen_name']
-
-        with self.assertLogs('respondent-home', 'INFO') as cm:
-            response = await self.client.request('POST',
-                                                 self.post_webchat_ni,
-                                                 data=form_data)
-        self.assertLogEvent(cm, 'form submission error')
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.nisra_logo, contents)
-        self.assertMessagePanel(WEBCHAT_MISSING_NAME_MSG, contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_open_en(self):
-        mocked_now = datetime.datetime(2019, 6, 15, 9, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            response = await self.client.request('POST',
-                                                 self.post_webchat_en,
-                                                 allow_redirects=False,
-                                                 data=self.webchat_form_data)
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.ons_logo_en, contents)
-        self.assertIn('iframe', contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_open_cy(self):
-        mocked_now = datetime.datetime(2019, 6, 15, 9, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            response = await self.client.request('POST',
-                                                 self.post_webchat_cy,
-                                                 allow_redirects=False,
-                                                 data=self.webchat_form_data)
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.ons_logo_cy, contents)
-        self.assertIn('iframe', contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_open_ni(self):
-        mocked_now = datetime.datetime(2019, 6, 15, 9, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            response = await self.client.request('POST',
-                                                 self.post_webchat_ni,
-                                                 allow_redirects=False,
-                                                 data=self.webchat_form_data)
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.nisra_logo, contents)
-        self.assertIn('iframe', contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_not_open_200_en(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url, status=200)
-
-                response = await self.client.request(
-                    'POST',
-                    self.post_webchat_en,
-                    allow_redirects=False,
-                    data=self.webchat_form_data)
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.ons_logo_en, contents)
-        self.assertIn('Bank Holidays', contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_not_open_200_cy(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url, status=200)
-
-                response = await self.client.request(
-                    'POST',
-                    self.post_webchat_cy,
-                    allow_redirects=False,
-                    data=self.webchat_form_data)
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.ons_logo_cy, contents)
-        self.assertIn('Gwyliau Banc', contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_not_open_200_ni(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url, status=200)
-
-                response = await self.client.request(
-                    'POST',
-                    self.post_webchat_ni,
-                    allow_redirects=False,
-                    data=self.webchat_form_data)
-
-        self.assertEqual(response.status, 200)
-        contents = str(await response.content.read())
-        self.assertIn(self.nisra_logo, contents)
-        self.assertIn('Bank Holidays', contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_not_open_clientconnectionerror_en(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url,
-                           exception=ClientConnectionError('Failed'))
-
-                with self.assertLogs('respondent-home', 'ERROR') as cm:
-                    response = await self.client.request(
-                        'POST',
-                        self.post_webchat_en,
-                        allow_redirects=False,
-                        data=self.webchat_form_data)
-                self.assertLogEvent(cm, 'client failed to connect')
-                self.assertLogEvent(cm, 'failed to send webchat closed')
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.ons_logo_en, contents)
-            self.assertIn('Bank Holidays', contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_not_open_clientconnectionerror_cy(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url,
-                           exception=ClientConnectionError('Failed'))
-
-                with self.assertLogs('respondent-home', 'ERROR') as cm:
-                    response = await self.client.request(
-                        'POST',
-                        self.post_webchat_cy,
-                        allow_redirects=False,
-                        data=self.webchat_form_data)
-                self.assertLogEvent(cm, 'client failed to connect')
-                self.assertLogEvent(cm, 'failed to send webchat closed')
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.ons_logo_cy, contents)
-            self.assertIn('Gwyliau Banc', contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_not_open_clientconnectionerror_ni(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url,
-                           exception=ClientConnectionError('Failed'))
-
-                with self.assertLogs('respondent-home', 'ERROR') as cm:
-                    response = await self.client.request(
-                        'POST',
-                        self.post_webchat_ni,
-                        allow_redirects=False,
-                        data=self.webchat_form_data)
-                self.assertLogEvent(cm, 'client failed to connect')
-                self.assertLogEvent(cm, 'failed to send webchat closed')
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.nisra_logo, contents)
-            self.assertIn('Bank Holidays', contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_not_open_500_en(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url, status=500)
-
-                with self.assertLogs('respondent-home', 'ERROR') as cm:
-                    response = await self.client.request(
-                        'POST',
-                        self.post_webchat_en,
-                        allow_redirects=False,
-                        data=self.webchat_form_data)
-                self.assertLogEvent(cm, 'failed to send webchat closed')
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.ons_logo_en, contents)
-            self.assertIn('Bank Holidays', contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_not_open_500_cy(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url, status=500)
-
-                with self.assertLogs('respondent-home', 'ERROR') as cm:
-                    response = await self.client.request(
-                        'POST',
-                        self.post_webchat_cy,
-                        allow_redirects=False,
-                        data=self.webchat_form_data)
-                self.assertLogEvent(cm, 'failed to send webchat closed')
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.ons_logo_cy, contents)
-            self.assertIn('Gwyliau Banc', contents)
-
-    @unittest_run_loop
-    async def test_post_webchat_not_open_500_ni(self):
-        mocked_now = datetime.datetime(2019, 6, 16, 16, 30, 00, 0)
-        with mock.patch('app.handlers.WebChat.get_now') as mocked_get_now:
-            mocked_get_now.return_value = mocked_now
-
-            with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-                mocked.get(self.webchatsvc_url, status=500)
-
-                with self.assertLogs('respondent-home', 'ERROR') as cm:
-                    response = await self.client.request(
-                        'POST',
-                        self.post_webchat_ni,
-                        allow_redirects=False,
-                        data=self.webchat_form_data)
-                self.assertLogEvent(cm, 'failed to send webchat closed')
-
-            self.assertEqual(response.status, 200)
-            contents = str(await response.content.read())
-            self.assertIn(self.nisra_logo, contents)
-            self.assertIn('Bank Holidays', contents)
 
     def test_uac_hash_en(self):
         # Given some post data
